+++ src/ca/javajesus/game/entities/structures/CastleTower.java
@@ -15,15 +15,14 @@ public class CastleTower extends SolidEntity {
 	protected int color = Colors.get(-1, 111, Colors.fromHex("#8d8d8d"),
 			Colors.fromHex("#eefeff"));
 
-	public CastleTower(Level level, double x, double y) {
+	public CastleTower(Level level, int x, int y) {
 		super(level, x, y, 61, 175);
 		this.shadow = new Rectangle(width, (5 * height / 6));
-		this.shadow.setLocation((int) x + 12, (int) y);
+		this.shadow.setLocation(x + 12, y);
 		this.bounds = new Rectangle(width, (height / 6) - 8);
-		this.bounds.setLocation((int) x + 12, (int) y + shadow.height);
+		this.bounds.setLocation(x + 12, y + shadow.height);
 		level.addEntity(new Transporter(level, x + 41, y + 159,
-				new CastleInterior(new Point((int) x + 43, (int) y + 167),
-						this.level)));
+				new CastleInterior(new Point(x + 43, y + 167), this.level)));
 
 	}
