+++ src/ca/javajesus/game/entities/structures/Furniture.java
@@ -5,7 +5,7 @@ import ca.javajesus.level.Level;
 
 public class Furniture extends SolidEntity {
 
-	public Furniture(Level level, double x, double y, int width, int height) {
+	public Furniture(Level level, int x, int y, int width, int height) {
 		super(level, x, y, width, height);
 	}
