+++ src/ca/javajesus/game/entities/structures/SanCiscoSkyscraper.java
@@ -12,14 +12,15 @@ import ca.javajesus.level.RandomLevel;
 
 public class SanCiscoSkyscraper extends SolidEntity {
 
-	protected int color = Colors.get(-1, Colors.fromHex("#edece0"), Colors.fromHex("#ffffff"), Colors.fromHex("#030074"));
+	protected int color = Colors.get(-1, Colors.fromHex("#edece0"),
+			Colors.fromHex("#ffffff"), Colors.fromHex("#030074"));
 
-	public SanCiscoSkyscraper(Level level, double x, double y) {
+	public SanCiscoSkyscraper(Level level, int x, int y) {
 		super(level, x, y, 106, 338);
 		this.shadow = new Rectangle(width, (8 * height / 9));
-		this.shadow.setLocation((int) x, (int) y);
+		this.shadow.setLocation(x, y);
 		this.bounds = new Rectangle(width, (height / 9) - 8);
-		this.bounds.setLocation((int) x, (int) y + shadow.height);
+		this.bounds.setLocation(x, y + shadow.height);
 		level.addEntity(new TransporterGlass(level, x + 44, y + 660,
 				new RandomLevel(Game.WIDTH, Game.HEIGHT)));
 		level.addEntity(new TransporterGlass(level, x + 100, y + 660,
@@ -29,8 +30,8 @@ public class SanCiscoSkyscraper extends SolidEntity {
 	}
 
 	public void render(Screen screen) {
-		
-		screen.render((int) x, (int) y, color, Sprite.sanCisco_skyscraper);
+
+		screen.render(x, y, color, Sprite.sanCisco_skyscraper);
 
 	}
