+++ src/ca/javajesus/game/entities/structures/Transporter.java
@@ -10,20 +10,21 @@ import ca.javajesus.game.gfx.SpriteSheet;
 import ca.javajesus.level.Level;
 
 public class Transporter extends SolidEntity {
-	
+
 	public final Rectangle hitBox = new Rectangle(8, 16);
 	protected Level nextLevel;
 
-	public Transporter(Level currentLevel, double x, double y, Level nextLevel) {
+	public Transporter(Level currentLevel, int x, int y, Level nextLevel) {
 		super(currentLevel, x, y, 8, 16);
 		this.nextLevel = nextLevel;
-		this.hitBox.setLocation((int) x, (int) y); 
+		this.hitBox.setLocation(x, y);
 	}
-	
-	public Transporter(Level currentLevel, double x, double y, Level nextLevel, Point point) {
+
+	public Transporter(Level currentLevel, int x, int y, Level nextLevel,
+			Point point) {
 		super(currentLevel, x, y, 8, 16);
 		this.nextLevel = nextLevel;
-		this.hitBox.setLocation((int) x - 4, (int) y - 8); 
+		this.hitBox.setLocation(x - 4, y - 8);
 		nextLevel.spawnPoint = point;
 	}
 
@@ -36,11 +37,31 @@ public class Transporter extends SolidEntity {
 	}
 
 	public void render(Screen screen) {
-		
-		screen.render(x + 0, y + 0, 0 + 5 * 32, Colors.get(-1, 111, Colors.fromHex("#704200"), Colors.fromHex("#ffde00")), 0, 1, SpriteSheet.tiles);
-		screen.render(x + 8, y + 0, 1 + 5 * 32, Colors.get(-1, 111, Colors.fromHex("#704200"), Colors.fromHex("#ffde00")), 0, 1, SpriteSheet.tiles);
-		screen.render(x + 0, y + 8, 0 + 6 * 32, Colors.get(-1, 111, Colors.fromHex("#704200"), Colors.fromHex("#ffde00")), 0, 1, SpriteSheet.tiles);
-		screen.render(x + 8, y + 8, 1 + 6 * 32, Colors.get(-1, 111, Colors.fromHex("#704200"), Colors.fromHex("#ffde00")), 0, 1, SpriteSheet.tiles);
+
+		screen.render(
+				x + 0,
+				y + 0,
+				0 + 5 * 32,
+				Colors.get(-1, 111, Colors.fromHex("#704200"),
+						Colors.fromHex("#ffde00")), 0, 1, SpriteSheet.tiles);
+		screen.render(
+				x + 8,
+				y + 0,
+				1 + 5 * 32,
+				Colors.get(-1, 111, Colors.fromHex("#704200"),
+						Colors.fromHex("#ffde00")), 0, 1, SpriteSheet.tiles);
+		screen.render(
+				x + 0,
+				y + 8,
+				0 + 6 * 32,
+				Colors.get(-1, 111, Colors.fromHex("#704200"),
+						Colors.fromHex("#ffde00")), 0, 1, SpriteSheet.tiles);
+		screen.render(
+				x + 8,
+				y + 8,
+				1 + 6 * 32,
+				Colors.get(-1, 111, Colors.fromHex("#704200"),
+						Colors.fromHex("#ffde00")), 0, 1, SpriteSheet.tiles);
 	}
 
 }
